requirements.txt
fastapi
uvicorn
uuid
llama-cpp-python
pandas
sentence-transformers
chromadb
PyMuPDF